Move serializer into another class
