+++ src/main/java/com/alan199921/astral/serializing/AstralSerializers.java
@@ -1,0 +1,42 @@
+package com.alan199921.astral.serializing;
+
+import com.mojang.authlib.GameProfile;
+import net.minecraft.nbt.CompoundNBT;
+import net.minecraft.nbt.NBTUtil;
+import net.minecraft.network.PacketBuffer;
+import net.minecraft.network.datasync.DataSerializers;
+import net.minecraft.network.datasync.IDataSerializer;
+
+import javax.annotation.Nonnull;
+import java.util.Optional;
+
+public class AstralSerializers {
+    public static final IDataSerializer<Optional<GameProfile>> OPTIONAL_GAME_PROFILE = new IDataSerializer<Optional<GameProfile>>() {
+        @Override
+        public void write(@Nonnull PacketBuffer packetBuffer, @Nonnull Optional<GameProfile> gameProfile) {
+            if (gameProfile.isPresent()) {
+                packetBuffer.writeBoolean(true);
+                packetBuffer.writeCompoundTag(NBTUtil.writeGameProfile(new CompoundNBT(), gameProfile.get()));
+            }
+            else {
+                packetBuffer.writeBoolean(false);
+            }
+        }
+
+        @Override
+        @Nonnull
+        public Optional<GameProfile> read(@Nonnull PacketBuffer packetBuffer) {
+            return packetBuffer.readBoolean() ? Optional.of(NBTUtil.readGameProfile(packetBuffer.readCompoundTag())) : Optional.empty();
+        }
+
+        @Override
+        @Nonnull
+        public Optional<GameProfile> copyValue(@Nonnull Optional<GameProfile> gameProfile) {
+            return gameProfile;
+        }
+    };
+
+    static {
+        DataSerializers.registerSerializer(OPTIONAL_GAME_PROFILE);
+    }
+}
